+++ src/hooks/swr-hooks/post.swr-hooks.ts
@@ -1,8 +1,8 @@
 import useSWRInfinite from "swr/infinite";
 import useSWRMutation from "swr/mutation";
 
-import { API_ROUTES } from "@/ig-sdk/api-routes";
-import { apiSDKInstance } from "@/ig-sdk/ig-sdk.instance";
+import { API_ROUTES } from "@/api-sdk/api-routes";
+import { apiSDKInstance } from "@/api-sdk/api-sdk.instance";
 import {
   PostInterface,
   PostResponseInterface,
