+++ src/hooks/swr-hooks/user.swr-hooks.ts
@@ -1,7 +1,7 @@
 import useSWR from "swr";
 
-import { apiSDKInstance } from "@/ig-sdk/ig-sdk.instance";
-import { API_ROUTES } from "@/ig-sdk/api-routes";
+import { apiSDKInstance } from "@/api-sdk/api-sdk.instance";
+import { API_ROUTES } from "@/api-sdk/api-routes";
 
 export const useSWRGetDecodedUserToken = () => {
   const { data, error, isLoading, mutate } = useSWR(
