+++ tsconfig.json
@@ -28,7 +28,7 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "src/ig-sdk/http-client.api-client.ts"
+    "src/api-sdk/http-client.api-client.ts"
   ],
   "exclude": ["node_modules"]
 }
